% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchcumhaz.R
\name{pchcumhaz}
\alias{pchcumhaz}
\title{Computation of the cumulative hazard function in the PCH model}
\usage{
pchcumhaz(seqtime, cuts, alpha)
}
\arguments{
\item{seqtime}{the time sequence.}

\item{cuts}{the sequence of cuts.}

\item{alpha}{the value of the hazard function between each cut. Should be of length equal to length(cuts)+1.}
}
\description{
Given a sequence of time, a set of cuts and hazard values, compute the cumulative hazard function from a piecewise-constant hazard model
}
\examples{
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
seqtime=seq(0,100,by=10)
pchcumhaz(seqtime,cuts,alpha)
}
