% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmst.R
\name{Rmst}
\alias{Rmst}
\title{Restricted Mean Survival Time}
\usage{
Rmst(Time, status, tau)
}
\arguments{
\item{Time}{a continuous time variable.}

\item{status}{an indicator for censoring for the corresponding Time variable.}

\item{tau}{the time endpoint for computing the RMST}
}
\value{
returns the RMST along with the value \code{tau} that was used.
}
\description{
Compute the Restricted Mean Survival Time (RMST) for right-censored data
}
\seealso{
\code{\link{pseudoKM}}, \code{\link{RmstIC}}.
}
