% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmstIC.R
\name{RmstIC}
\alias{RmstIC}
\title{Restricted Mean Survival Time for Interval-Censored data}
\usage{
RmstIC(cuts, alpha, tau)
}
\arguments{
\item{cuts}{the cuts used in the PCH model.}

\item{alpha}{the value of the hazard function between each cut. Should be of
length equal to \code{length(cuts)}+1.}

\item{tau}{the time endpoint for computing the RMST.}
}
\description{
Compute the Restricted Mean Survival Time (RMST) for interval-censored data
from the piecewise constant hazard (PCH) model
}
