% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv}
\alias{rsurv}
\title{Simulation of survival data following a PCH model}
\usage{
rsurv(n, cuts = NULL, alpha)
}
\arguments{
\item{n}{the sample size.}

\item{cuts}{the sequence of cuts. Default to NULL which corresponds to the exponential model.}

\item{alpha}{the value of the hazard function between each cut. Should be of length equal to length(cuts)+1.}
}
\value{
A time vector of size \code{n} drawn from a piecewise constant hazard model with hazard equal to \code{alpha} between \code{cuts}.
}
\description{
Given a set of cuts and hazard values, generates a survival data sample from a piecewise-constant hazard model
}
\examples{
n=400
cuts=c(20,40,50,70)
alpha=c(0,0.05,0.1,0.2,0.4)/10
time=rsurv(n,cuts,alpha) #generate true data from the pch model
censoring=stats::runif(n,min=70,max=90)
time=pmin(time,censoring) #observed times
delta=time<censoring #gives 62\% of observed data on average
}
